+++ lib/src/Benchmark/bench_context.dart
@@ -1,0 +1,13 @@
+part of softhai.spec_dart;
+
+abstract class BenchContext {
+  
+  Map<String, dynamic> data;
+  
+}
+
+class _BenchContextImpl implements BenchContext {
+  
+  Map<String, dynamic> data = new Map<String, dynamic>();
+  
+}